+++ src/Contract/src/Handler/CC/PostCreateCCHandler.php
@@ -5,15 +5,17 @@ declare(strict_types=1);
 namespace Frontend\Contract\Handler\CC;
 
 use Core\App\Message;
+use Frontend\App\Common\Department;
+use Core\Contract\Enum\QueueFolderEnum;
 use Dot\DependencyInjection\Attribute\Inject;
 use Dot\FlashMessenger\FlashMessengerInterface;
 use Dot\Log\Logger;
 use Fig\Http\Message\StatusCodeInterface;
 use Frontend\App\Exception\ConflictException;
 use Frontend\Contract\Form\CreateCCForm;
-use Frontend\Contract\Service\CCServiceInterface;
-use Laminas\Diactoros\Response\EmptyResponse;
+use Frontend\Contract\Service\ContractServiceInterface;
 use Laminas\Diactoros\Response\HtmlResponse;
+use Laminas\Diactoros\Response\RedirectResponse;
 use Mezzio\Router\RouterInterface;
 use Mezzio\Template\TemplateRendererInterface;
 use Psr\Http\Message\ResponseInterface;
@@ -24,7 +26,7 @@ use Throwable;
 class PostCreateCCHandler implements RequestHandlerInterface
 {
     #[Inject(
-        CCServiceInterface::class,
+        ContractServiceInterface::class,
         RouterInterface::class,
         TemplateRendererInterface::class,
         FlashMessengerInterface::class,
@@ -32,7 +34,7 @@ class PostCreateCCHandler implements RequestHandlerInterface
         'dot-log.default_logger',
     )]
     public function __construct(
-        protected CCServiceInterface $cCService,
+        protected ContractServiceInterface $contractService,
         protected RouterInterface $router,
         protected TemplateRendererInterface $template,
         protected FlashMessengerInterface $messenger,
@@ -45,28 +47,60 @@ class PostCreateCCHandler implements RequestHandlerInterface
         ServerRequestInterface $request,
     ): ResponseInterface {
         $this->createCCForm
-            ->setAttribute('action', $this->router->generateUri('c-c::create-c-c'));
+            ->setAttribute('action', $this->router->generateUri('cc::post-create'));
 
         try {
             $data = (array) $request->getParsedBody();
             $this->createCCForm->setData($data);
             if ($this->createCCForm->isValid()) {
                 $data = $this->createCCForm->getData();
-                $this->cCService->saveCC($data);
-                $this->messenger->addSuccess(Message::C_C_CREATED);
-
-                return new EmptyResponse(StatusCodeInterface::STATUS_CREATED);
+                
+                $data['DOCTYPE'] = 'Contract';
+                $data['CONTRACT_AMOUNT'] = '';
+                $data['CONTRACT_END_DATE'] = '';
+                
+                //-- ExCESS_100K --//
+                if ($data['EXCESS_100K']) {
+                    $data['EXCESS_100K_Y'] = mb_chr(0x2611, 'UTF-8');
+                    $data['EXCESS_100K_N'] = mb_chr(0x2610, 'UTF-8');
+                } else {
+                    $data['EXCESS_100K_Y'] = mb_chr(0x2610, 'UTF-8');
+                    $data['EXCESS_100K_N'] = mb_chr(0x2611, 'UTF-8');
+                }
+                
+                //-- GOV_USE --//
+                if ($data['GOV_USE']) {
+                    $data['GOV_USE_Y'] = mb_chr(0x2611, 'UTF-8');
+                    $data['GOV_USE_N'] = mb_chr(0x2610, 'UTF-8');
+                } else {
+                    $data['GOV_USE_Y'] = mb_chr(0x2610, 'UTF-8');
+                    $data['GOV_USE_N'] = mb_chr(0x2611, 'UTF-8');
+                }
+                
+                $a = QueueFolderEnum::from($data['DEPARTMENT'])->name;
+                $ref = new \ReflectionClass(Department::class);
+                $data['DEPARTMENT_NAME'] = $ref->getConstant($a);
+                
+                /**
+                 * Create Contract
+                 */
+                $contract = $this->contractService->createContract($data);
+                $this->contractService->generateContract($data, $contract);
+                
+                $this->messenger->addSuccess('Contract Created');
+                
+                return new RedirectResponse($this->router->generateUri('workflow::dashboard', ['action' => 'index']));
             }
 
             return new HtmlResponse(
-                $this->template->render('c-c::create-c-c-form', [
+                $this->template->render('contract::create-contract-form', [
                     'form' => $this->createCCForm->prepare(),
                 ]),
                 StatusCodeInterface::STATUS_UNPROCESSABLE_ENTITY
             );
         } catch (ConflictException $exception) {
             return new HtmlResponse(
-                $this->template->render('c-c::create-c-c-form', [
+                $this->template->render('contract::create-contract-form', [
                     'form'     => $this->createCCForm->prepare(),
                     'messages' => [
                         'error' => $exception->getMessage(),
@@ -83,7 +117,7 @@ class PostCreateCCHandler implements RequestHandlerInterface
             ]);
 
             return new HtmlResponse(
-                $this->template->render('c-c::create-c-c-form', [
+                $this->template->render('contract::create-contract-form', [
                     'form'     => $this->createCCForm->prepare(),
                     'messages' => [
                         'error' => Message::AN_ERROR_OCCURRED,